bootstrap: made a better generic build rule
The build script's default case now has a decent rule.  A bunch of variables
are set at the beginning of the function that simplify all the other rules,
too.

This commit also adds a binutils rule.  zlib and binutils have not been
tested yet, but everything prior to zlib works (at least it did the one time
I tried it...)