bootstrap: added fix-perms
This commit fixed a bug in e2fsprogs script and adds fix-perms.  I'm not
sure I like the way fix-perms is implemented at the moment (su -c and ask
for root pw), but this works for now.

At this point, the bootstrap areas can completely build a usable root
filesystem.  Code still needs to be added (maybe) to create a partition and
configure grub... but I might just leave this up to the user for now.