kernel: added src/kernel submodule
This is currently pointing at the HEAD of stable-2.6.37.y.

Note that I did have a commit on my old 2.6.36-mdl kernel that helped the
build system find ncurses when it's not in /usr... I might have to cherry-
pick this later, but it should only be needed during the bootstrap stage
(when we have no control over where ncurses lives).  This might be
annoying, though, because that's when the initial configuring of the
bootstrap kernel has to happen...