bootstrap: trying to fix autoreconf cross-compiler problems, v2
This commit adds set_cross_compile and unset_cross_compile functions, and
modifies autoreconf_generic to save and reset cross-compilation state.  In
other words, autoreconf_generic makes sure that a) cross-compiler settings
are NOT used when bootstrapping a source tree and b) cross-compiler
settings are preserved unchanged in the parent function.

This also adds a block of code to setup_generic to apply patches specified
by the patches variable prior to calling autoreconf_generic.

This patch also fixes some typos, cleans up some comments, and undoes most
of what was done in v1.