reorg: now expect source tree to be inside our baby filesystem
In order for bootstrap-stage2 (post-boot) and beyond to have access to the
ruckus source tree, it has to be inside the baby system's rootfs.  Duh.
Slight oversight during the initial design, fixed now.

This commit introduces a new variable in functions, BUILDROOT, and tweaks
the existing variables so that everything is being built/installed
temporarily inside $BUILDROOT (which is inside ruckus source tree), then
the important stuff gets moved to $INSTALLROOT (which should be the
toplevel dir of the baby rootfs that the ruckus source tree was cloned
into).

The end result of this is a bootable rootfs (the result of bootstrap-
stage1) that has the source tree we've been working in all along already
inside it at /ruckus.  After booting, bootstrap-stage2 can pick up where
bootstrap-stage1 left off and finish building all the development tools
we'll need to build the final system!

The next step is to actually create bootstrap-stage2 and cause the
bootscripts setup in bootstrap-stage1 to automatically kick it off after
booting!