bootstrap: implemented some functions for ash
This commit adds a bunch of definitions to functions to make up for
bash-isms that aren't in ash (e.g., pushd/popd).

Note that the effect this has on previous bootsrap stages that actually use
bash has not yet been worked out.

Alos note that the 'sudo' function added in this commit doesn't quite work
right yet.