gitmodules: use real remote git urls, v2
This commit fixes a couple submodules that were now missing the commit we
were checking out before.  The change for srp is because I switched from
the v2 repo to the v3 repo in the previous commit.  The other modules were
because I had been using home made git repos, which I still need to go sort
out (if any of my old patches are still needed).