bootstrap-early: don't force autoreconf
This shouldn't really be needed.  If configure is missing, the autoreconf
script of choice will automagically get invoked.  If configure is already
there, why waste our time?

I think I had this in there originally just to make sure I worked the kinks
out of the bootstrapping process, but it should no longer be needed.