bootstrap: consistently use CLFS variables
This commit brings us back in common with the variables used in the CLFS
book.  I had changed a bunch of them in the initial implementation because
I felt like it... but that was probably not a good idea.  This way we can
easily cut-n-paste scripts from the CLFS book.