get_sources: iterate over results of find
Before we were doing a bunch of shell globs, based on where I knew we had
SOURCE_URL files.  This will be more dynamic.  For example, now I can add a
SOURCE_URL file inside a submodule in src, and as long as I've updated my
submods first, I'll be able to run get_sources to autodownload the needed
files.