gnulib: updated to head of master
This may be involved in the m4 build failure mentioned previously.  When we
updated m4, I noticed that the gnulib submodule wasn't behaving correctly,
but I forgot all about it.  We need to keep our src/gnulib submod as up to
date as possible, because our init_submodules script uses it as a shared
repo for all the other gnulib submods in the system...  m4 was referencing
a sha that was newer than what we had available in src/gnulib, so it was
getting some wicked old checkout...