autotools: playing with build rules for packages, v2
I've got two examples now, ruckusrd and syslinux.  The former uses a git
submoodule, the latter a fetched source tarball.

Neither Makefile.am has any reference to the name of the package, it's
source tarball, version number, etc.  Everything is being determined
dynamically by make.

The only difference between the two automake files is the additional rules
for bootstrapping and making dist needed if we're not using a fetched dist
tarball.  Which means if we conditionally include those rules, we're
golden.