bootstrap-stage2: fixed login shell corruption
This turned out to be related to running a getty on /dev/ttyS0 while logged
in via Xen's /dev/hvc0...

This commit reworks the installed inittab file (mostly just comment
cleanup) and turns off the ttyS0 login.  It also replaces the wonky bash
invocation in the stage2 init script with a simple login prompt.