bootstrap-stage1: fix XEN_TARGET_ARCH for 32bit builds
This variable has to be set to x86_32, not i386, for 32bit builds.  With
the variable set to match CLFS_ARCH (i386), the xen build fails with an
error about missing a makefile.

That being said, when correctly configured to build for 32bit, Xen issues
a warning about no longer supporting 32bit builds and skips it.  So it's
almost silly.