bootstrap: support parallel out-of-tree builds, v1
The 1st problem we need to overcome is the fact that all builds use /tools
and /cross-tools symlinks.  Proposed solution is appending
the CLFS_ARCH_STRING (e.g., x86_64-64) to /tools and /cross-tools.  This
way each build will have it's own symlinks on the host system and they
won't clobber each other.

This commit just decorates all the places I need to tweak with FIXMEs.  In
addition, we will have to sed |tools|tools-archstring| in the gcc patches
in both bootstrap-cross and bootstrap-stage1.