bootstrap: use CLFS-tested kernel_headers but newer actual kernel
The idea here is to use the exact kernel_header/glibc combo that CLFS is
using regardless of what tag I have checked out in src/kernel.  This will
allow me to use a much newer kernel than CLFS for booting and running the
resulting system, but not have to worry about borking glibc.