autoconf doesn't like newer versions of texinfo
I didn't see this on smaug because texinfo is provided by avlinux (so we
don't build it in hostprep) but old (so autoconf builds fine).

Gonna have to patch autoconf.texi, otherwise final package won't be able
to include the docs... which would suck.

Debian Bug 711297 has a patch, which looks mostly like a line-wrapping
change...