bootstrap-stage1: updated build scripts for CLFS GIT-20170803, v4
Finished 64bit kernel config update.  So, this isn't a small fiesty
bootstrap kernel like the CLFS book suggest.  I've gone and made this a
large distro-ish kernel with tons of drivers so that I can use it as the
kernel for my CentOS 7 host.  This gives me a pvops Xen kernel for the host
system as well as a (hopefully) usable kernel for the bootstrap guest.

NOTE: i disabled /sbin/hotplug and enabled devtmpfs instead... hope it
      works when booting bootstrap system...

NOTE: i've got a bunch of security stuff turned on to mimic CentOS 7's
      kernel config, including SELinux... hope it works ok.