upgrade to GCC 8.2 and GLIBC 2.28 for kernel's CONFIG_RETPOLINE
So, I probably could have just upgraded to GCC 7.4 (or even just 7.3) to
turn CONFIG_RETPOLINE on in the kernel configs... but I figured, hey, why
not upgrade to the latest and greatest while we're at it?

Upgrading to GCC 8.2 resulting in compilation failures in glibc because the
new compiler finds more warnings and glibc builds w/ -Werror.  So, I've
gone and upgraded glibc to the latest version as well.

As such, I've also also updated the installed kernel headers to match what
was the longterm kernel when that version of glibc was released.

I also updated to final kernel to v4.19.15.

I also ran into problems with glibc using host system's gmake instead of
hostprep's make... which of course was too old.  So I added symlinks in
bootstrap-early's make scriptlet.

I have not updated gmp, mpfr, mpc, or isl... so this might go boom.